backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
sse-starlette==1.8.2
